refactor: rename can_live→tm3web, tm3diag.py→tm3cli.py; trim docs
- can_live.py → tm3web.py, can_live_ui/ → tm3web_ui/ (it's a local web
  console now: Live / Alerts / DB Explorer / Raw / ODIN / DID)
- tm3diag.py → tm3cli.py (distinguish the CLI from the project name);
  not imported as a module, so no code-import churn
- docs: ghidra_c28x_loading cut to 4 steps + link the upstream setup guide;
  di.md immobilizer via the provider seam; tm3web.md console doc;
  bhx dup heading; FIRMWARE_UPDATE tools → bhx.py; unsquash xdis → pycdc

No behavior change.
